Refactor: move comparison logic back to the `equatable_utils.dart` file
